Добавил возможность записи media файлов и даты на сервер через yml config.
