+++ Sample/models.py
@@ -20,7 +20,8 @@ class AbstractModel(models.Model):
 
 
 class SampleType(AbstractModel):
-    name = models.CharField(max_length=255, blank=True, null=True)
+    """Тип образца"""
+    name = models.CharField(max_length=255, verbose_name='Название типа образца', default='')
 
     class Meta:
         db_table = 'sample_type'
@@ -29,13 +30,15 @@ class SampleType(AbstractModel):
         return f'{self.name}'
 
     # Добавил для AjaxableResponseMixin, для реализции DetailView необходиму возращать url элемента
-    def get_absolute_url(self):
+    @staticmethod
+    def get_absolute_url():
         return reverse('sample:type_table')
 
 
 class SampleMaterial(AbstractModel):
-    name = models.CharField(max_length=255, blank=True, null=True)
-    type = models.CharField(max_length=255, blank=True, null=True)
+    """Материал из которого изготовлен образец"""
+    name = models.CharField(max_length=255, verbose_name='Название материала образца', default='')
+    type = models.CharField(max_length=255, blank=True, null=True, verbose_name='Тип материала')
     description = models.TextField(blank=True, null=True)
 
     class Meta:
@@ -50,33 +53,36 @@ class SampleMaterial(AbstractModel):
             label += self.type
         return label
 
-    def get_absolute_url(self):
+    @staticmethod
+    def get_absolute_url():
         return reverse('sample:material_table')
 
 
 class Sample(AbstractModel):
-
+    """Информация по образцу"""
     def sample_image_path(self, filename):
         return f"sample/{self.id}/{filename}"
 
-    method = models.CharField(max_length=255, blank=True, null=True)
-    exp_param = models.CharField(max_length=255, blank=True, null=True)
-    date_proc_streng = models.DateField(blank=True, null=True)
-    depth_coating = models.FloatField(blank=True, null=True)
-    roughness = models.FloatField(blank=True, null=True)
-    hardness_coating = models.FloatField(blank=True, null=True)
-    struct_coating = models.CharField(max_length=255, blank=True, null=True)
-    sub_hardness = models.FloatField(blank=True, null=True)
-    organization = models.CharField(max_length=255, blank=True, null=True)
-    weight = models.FloatField(blank=True, null=True)
-    marking = models.CharField(max_length=255, blank=True, null=True)
-    image = models.FileField(upload_to=sample_image_path, null=True, blank=True)
+    method = models.CharField(max_length=255, blank=True, null=True, verbose_name='Способ упрочнения')
+    exp_param = models.CharField(max_length=255, blank=True, null=True, verbose_name='Параметры упрочнения')
+    date_proc_streng = models.DateField(blank=True, null=True, verbose_name='Дата процесса упрочнения')
+    depth_coating = models.FloatField(blank=True, null=True, verbose_name='Толщина упрочнения')
+    roughness = models.FloatField(blank=True, null=True, verbose_name='Шероховатость поверхности')
+    hardness_coating = models.FloatField(blank=True, null=True, verbose_name='Твердость упрочнения')
+    struct_coating = models.CharField(max_length=255, blank=True, null=True, verbose_name='Состав покрытия')
+    sub_hardness = models.FloatField(blank=True, null=True, verbose_name='Твердость подложки')
+    organization = models.CharField(max_length=255, blank=True, null=True,
+                                    verbose_name='Организация, которая провела упрочнение')
+    weight = models.FloatField(verbose_name='Первоначальная масса образца', default=0)
+    marking = models.CharField(max_length=255, blank=True, null=True, verbose_name='Маркировка образца')
+    image = models.FileField(upload_to=sample_image_path, null=True, blank=True,
+                             verbose_name='Внешний вид образца, начальный')
     description = models.TextField(blank=True, null=True)
 
-    sample_material = models.ForeignKey(SampleMaterial, on_delete=models.CASCADE,
-                                        blank=True, null=True, related_name='material')
-    sample_type = models.ForeignKey(SampleType, on_delete=models.CASCADE,
-                                    blank=True, null=True, related_name='type')
+    sample_material = models.ForeignKey(SampleMaterial, blank=True, null=True,
+                                        on_delete=models.CASCADE, related_name='material')
+    sample_type = models.ForeignKey(SampleType, blank=True, null=True,
+                                    on_delete=models.CASCADE, related_name='type')
 
     class Meta:
         db_table = 'sample'
@@ -101,17 +107,17 @@ class Sample(AbstractModel):
 
 
 class Trials(AbstractModel):
-    size_particle = models.IntegerField(blank=True, null=True)
-    speed_collision = models.IntegerField(blank=True, null=True)
-    add_param = models.CharField(max_length=255, blank=True, null=True)
-    corner_collision = models.IntegerField(blank=True, null=True)
-    date_trials = models.DateField(blank=True, null=True)
-    date_end_trials = models.DateField(blank=True, null=True)
-    type_particle = models.CharField(max_length=255, blank=True, null=True)
+    """Информация по испытанию"""
+    size_particle = models.IntegerField(blank=True, null=True, verbose_name='Размер частиц')
+    speed_collision = models.IntegerField(blank=True, null=True, verbose_name='Скорость соударения')
+    add_param = models.CharField(max_length=255, blank=True, null=True, verbose_name='Дополнительные параметры частицы')
+    corner_collision = models.IntegerField(blank=True, null=True, verbose_name='Угол соударения ')
+    date_trials = models.DateField(blank=True, null=True, verbose_name='Дата проведения испытания')
+    date_end_trials = models.DateField(blank=True, null=True, verbose_name='Время окончания испытания')
+    type_particle = models.CharField(max_length=255, blank=True, null=True, verbose_name='Тип частицы')
     description = models.TextField(blank=True, null=True)
 
-    sample = models.OneToOneField(Sample, on_delete=models.CASCADE,
-                                  blank=True, null=True, related_name='sample')
+    sample = models.OneToOneField(Sample, blank=True, null=True, on_delete=models.CASCADE, related_name='sample')
 
     class Meta:
         db_table = 'trials'
@@ -121,12 +127,14 @@ class Trials(AbstractModel):
 
 
 class ReceivedValues(AbstractModel):
+    """Информация по эксперементу"""
     def experement_image_path(self, filename):
         return f"experement/trial_{self.trials.id}/{filename}"
 
-    change_weight = models.FloatField(blank=True, null=True)
-    time_trials = models.IntegerField(blank=True, null=True)
-    image = models.FileField(upload_to=experement_image_path, null=True, blank=True)
+    change_weight = models.FloatField(verbose_name='Масса образца после эксперимента', default=0)
+    time_trials = models.IntegerField(verbose_name='Длительность эксперимента', default=0)
+    image = models.FileField(upload_to=experement_image_path, null=True, blank=True,
+                             verbose_name='Изображение образца после эксперимента')
 
     trials = models.ForeignKey(Trials, on_delete=models.CASCADE,
                                blank=True, null=True, related_name="trials_values")
@@ -134,5 +142,6 @@ class ReceivedValues(AbstractModel):
     class Meta:
         db_table = 'received_values'
 
-    def get_absolute_url(self):
+    @staticmethod
+    def get_absolute_url():
         return reverse('trial:list')
