+++ Sample/templates/sample/sample.html
@@ -157,7 +157,6 @@ $('.ui.checkbox.all')
     }
 
     function ajax_get_graph(trial_id, sample_id){
-        // Реализовать проверку если trial_id отсутствует
         $.ajax({
             type: 'GET',
             url: '/trial/' + trial_id + '/get_graph',
@@ -172,6 +171,7 @@ $('.ui.checkbox.all')
             type: 'GET',
             url: '/sample/detail/' + sample_id,
             success: function (response) {
+                $('#id_sample_modal').remove(); // Удаляем модальное окно из кеша
                 let modal = $(response);
                 modal.find('#id_graph').append(graph);
                 modal.modal('show');
