+++ Sample/templates/sample/sample_modal_detail.html
@@ -1,4 +1,4 @@
-<div class="ui modal sample">
+<div class="ui modal sample" id="id_sample_modal">
     <i class="close icon"></i>
     <div class="header">
         {% if sample.sample_material.name %}
